Change input to double

+++ src/java/syafa/java/Praktikum/Bab3/Praktikum2.java
@@ -15,7 +15,7 @@ public class Praktikum2
                             "4. HITUNG VOLUME SILINDER",
                             "5. HITUNG VOLUME LIMAS SEGITIGA",
                             "MASUKKAN PILIHAN ANDA",':');
-        int selection = Integer.parseInt(input.nextLine());
+        double selection = Double.parseDouble(input.nextLine());
         input.close();
         boolean inputValid = false;
         while (!inputValid)
